feat: Add QuizQuestion model and update related entities for quiz management

+++ 06-06-2025/QuizupAPI/Models/Answer.cs
@@ -3,9 +3,9 @@ namespace QuizupAPI.Models
     public class Answer
     {
         public long Id { get; set; }
-        public int QuestionId { get; set; }
+        public long QuestionId { get; set; }
         public Question? Question { get; set; }
-        public int QuizSubmissionId { get; set; }
+        public long QuizSubmissionId { get; set; }
         public QuizSubmission? QuizSubmission { get; set; }
         public string SelectedAnswer { get; set; } = string.Empty;
     }
