+++ 06-06-2025/QuizupAPI/Models/QuizSubmission.cs
@@ -3,13 +3,13 @@ namespace QuizupAPI.Models
     public class QuizSubmission
     {
         public long Id { get; set; }
-        public int QuizId { get; set; }
+        public long QuizId { get; set; }
         public Quiz? Quiz { get; set; }
-        public int StudentId { get; set; }
+        public long StudentId { get; set; }
         public Student? Student { get; set; }
         public DateTime SubmissionDate { get; set; } = DateTime.Now;
         public List<Answer>? Answers { get; set; }
         public string SubmissionStatus { get; set; } = "InProgress";
-        public int? Score { get; set; }
+        public long? Score { get; set; }
     }
 }
